+++ src/features/tracking-sidebar/ui/TrackingSidebar.tsx
@@ -1,0 +1,5 @@
+import { Box } from '@/shared/ui/Box'
+import * as style from './tracking-sidebar.css'
+export const TrackingSidebar = () => {
+    return <Box className={style.layout}></Box>
+}
