+++ src/features/tracking-sidebar/ui/tracking-sidebar.css.ts
@@ -1,0 +1,10 @@
+import { colors } from '@/app/token'
+import { style } from '@vanilla-extract/css'
+
+export const layout = style({
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    flexDirection: 'row',
+    backgroundColor: `${colors['neutral-10']}`,
+})
