+++ src/pages/tracking/Tracking.tsx
@@ -1,10 +1,19 @@
 import { Box } from '@/shared/ui/Box'
-import { TrackingSidebar } from '@/widgets/tracking-sidebar/ui/TrackingSidebar'
-import * as style from './tracking.css'
+import { TrackingSidebar } from '@/features/tracking-sidebar/ui/TrackingSidebar'
+import { Sidebar } from '@/widgets/Sidebar'
+import { colors } from '@/app/token'
 export const Tracking = () => {
     return (
-        <Box className={style.layout}>
-            <TrackingSidebar />
+        <Box
+            display="flex"
+            alignItems="center"
+            justifyContent="center"
+            flexDirection="row"
+            style={{ gap: '10px', backgroundColor: `${colors['neutral-10']}` }}
+        >
+            <Sidebar headerText="추적">
+                <TrackingSidebar />
+            </Sidebar>
         </Box>
     )
 }
