+++ src/widgets/Sidebar/Sidebar-Footer/index.ts
