+++ src/widgets/Sidebar/Sidebar-Footer/sidebar-footer.css.ts
