--- src/widgets/tracking-sidebar/ui/tracking-sidebar.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css'
-
-export const layout = style({
-    position: 'relative',
-    top: '30px',
-    marginBottom: '37px',
-})
